# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "App developement Organization - MashUp : FoodFeed Backend API Swagger Document\n",
    "version" : "1.0.0",
    "title" : "FoodFeed",
    "contact" : {
      "email" : "ueong777@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/woozero-k/foodfeed/1.0.0",
  "tags" : [ {
    "name" : "auth",
    "description" : "register, login, nickname_check"
  }, {
    "name" : "feed",
    "description" : "write, delete, modify, list, detail, like, unlike, is_like,"
  }, {
    "name" : "search",
    "description" : "search"
  }, {
    "name" : "user",
    "description" : "info_detail, user_feeds, follow, unfollow, modify"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/auth/facebook" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "register via facebook",
        "description" : "ok",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "register",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Auth"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "register successfully"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    },
    "/feed" : {
      "post" : {
        "tags" : [ "feed" ],
        "summary" : "register via facebook",
        "description" : "ok",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "register",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Auth"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "register successfully"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    },
    "/feed/:feed_id" : {
      "put" : {
        "tags" : [ "feed" ],
        "summary" : "register via facebook",
        "description" : "ok",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "register",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Auth"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "register successfully"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    },
    "/feed/:feed_id" : {
      "delete" : {
        "tags" : [ "feed" ],
        "summary" : "register via facebook",
        "description" : "ok",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "register",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Auth"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "register successfully"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    }
  },
  "definitions" : {
    "Auth" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "petId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "shipDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "status" : {
          "type" : "string",
          "description" : "Order Status",
          "enum" : [ "placed", "approved", "delivered" ]
        },
        "complete" : {
          "type" : "boolean",
          "default" : false
        }
      },
      "example" : {
        "petId" : 6,
        "quantity" : 1,
        "id" : 0,
        "shipDate" : "2000-01-23T04:56:07.000+00:00",
        "complete" : false,
        "status" : "placed"
      }
    }
  },
  "externalDocs" : {
    "description" : "Find out more FoodFeed",
    "url" : "https://noverish.github.io/FoodFeedDoc/#api-auth-auth_login"
  }
}